01-report-narrator/requirements.txt
anthropic>=0.40.0
streamlit>=1.35.0
pandas>=2.0.0
openpyxl>=3.1.0
jinja2>=3.1.0
python-dotenv>=1.0.0
numpy>=1.26.0
